DMP-1128 Transcription state model restrictions
Made changes as per code review
