+++ src/main/java/uk/gov/hmcts/darts/transcriptions/exception/TranscriptionApiError.java
@@ -28,8 +28,13 @@ public enum TranscriptionApiError implements DartsApiError {
         HttpStatus.BAD_REQUEST,
         "The workflow comment is required for this transcription update"
     ),
-    TRANSCRIPTION_WORKFLOW_ACTION_INVALID(
+    BAD_REQUEST_TRANSCRIPTION_TYPE(
         "104",
+        HttpStatus.BAD_REQUEST,
+        "Unexpected transcription type for this workflow"
+    ),
+    TRANSCRIPTION_WORKFLOW_ACTION_INVALID(
+        "105",
         HttpStatus.CONFLICT,
         "Transcription workflow action is not permitted"
     );
