+++ src/main/java/uk/gov/hmcts/darts/transcriptions/validator/WorkflowValidator.java
@@ -6,7 +6,7 @@ import uk.gov.hmcts.darts.transcriptions.enums.TranscriptionStatusEnum;
 import uk.gov.hmcts.darts.transcriptions.enums.TranscriptionTypeEnum;
 
 import java.util.Collections;
-import java.util.HashMap;
+import java.util.EnumMap;
 import java.util.Map;
 import java.util.Set;
 
@@ -23,13 +23,14 @@ import static uk.gov.hmcts.darts.transcriptions.enums.TranscriptionTypeEnum.OTHE
 @Slf4j
 public class WorkflowValidator {
 
-    // Create a map defining the transition rules, from any given current state to a list of allowable target states
-    private final Map<TranscriptionStatusEnum, Set<TranscriptionStatusEnum>> manualWorkflowTransitionRules = new HashMap<>();
+    private final Map<TranscriptionStatusEnum, Set<TranscriptionStatusEnum>> manualWorkflowTransitionRules =
+        new EnumMap<>(TranscriptionStatusEnum.class);
 
-    private final Map<TranscriptionStatusEnum, Set<TranscriptionStatusEnum>> automaticWorkflowTransitionRules = new HashMap<>();
+    private final Map<TranscriptionStatusEnum, Set<TranscriptionStatusEnum>> automaticWorkflowTransitionRules =
+        new EnumMap<>(TranscriptionStatusEnum.class);
 
     public WorkflowValidator() {
-        manualWorkflowTransitionRules.put(REQUESTED, Set.of(REQUESTED, AWAITING_AUTHORISATION, CLOSED));
+        manualWorkflowTransitionRules.put(REQUESTED, Set.of(AWAITING_AUTHORISATION, CLOSED));
         manualWorkflowTransitionRules.put(AWAITING_AUTHORISATION, Set.of(APPROVED, REJECTED, CLOSED));
         manualWorkflowTransitionRules.put(APPROVED, Set.of(WITH_TRANSCRIBER, CLOSED));
         manualWorkflowTransitionRules.put(REJECTED, Collections.emptySet());
@@ -37,7 +38,7 @@ public class WorkflowValidator {
         manualWorkflowTransitionRules.put(COMPLETE, Collections.emptySet());
         manualWorkflowTransitionRules.put(CLOSED, Collections.emptySet());
 
-        automaticWorkflowTransitionRules.put(REQUESTED, Set.of(REQUESTED, APPROVED, CLOSED));
+        automaticWorkflowTransitionRules.put(REQUESTED, Set.of(APPROVED, CLOSED));
         automaticWorkflowTransitionRules.put(AWAITING_AUTHORISATION, Collections.emptySet());
         automaticWorkflowTransitionRules.put(APPROVED, Set.of(WITH_TRANSCRIBER, CLOSED));
         automaticWorkflowTransitionRules.put(REJECTED, Collections.emptySet());
@@ -51,26 +52,26 @@ public class WorkflowValidator {
     }
 
     public boolean validateChangeToWorkflowStatus(TranscriptionTypeEnum transcriptionTypeEnum,
-                                               TranscriptionStatusEnum currentTranscriptionStatus,
-                                               TranscriptionStatusEnum desiredTargetTranscriptionStatus) {
-        boolean success = true;
+                                                  TranscriptionStatusEnum currentTranscriptionStatus,
+                                                  TranscriptionStatusEnum desiredTargetTranscriptionStatus) {
         if (isAutomatedTranscription(transcriptionTypeEnum)) {
-            if (!automaticWorkflowTransitionRules.get(currentTranscriptionStatus).contains(desiredTargetTranscriptionStatus)) {
-                success = logInvalidTranscriptionWorkflow(transcriptionTypeEnum, currentTranscriptionStatus, desiredTargetTranscriptionStatus);
-            }
-        } else if (!manualWorkflowTransitionRules.get(currentTranscriptionStatus).contains(desiredTargetTranscriptionStatus)) {
-            success = logInvalidTranscriptionWorkflow(transcriptionTypeEnum, currentTranscriptionStatus, desiredTargetTranscriptionStatus);
+            return isValid(automaticWorkflowTransitionRules, currentTranscriptionStatus, desiredTargetTranscriptionStatus, transcriptionTypeEnum);
         }
-        return success;
+        return isValid(manualWorkflowTransitionRules, currentTranscriptionStatus, desiredTargetTranscriptionStatus, transcriptionTypeEnum);
     }
 
-
-    private boolean logInvalidTranscriptionWorkflow(TranscriptionTypeEnum transcriptionTypeEnum,
-                                                    TranscriptionStatusEnum currentTranscriptionStatus,
-                                                    TranscriptionStatusEnum desiredTargetTranscriptionStatus) {
-
-        log.warn("Unable to go from workflow state {} to {} for type {}", currentTranscriptionStatus, desiredTargetTranscriptionStatus, transcriptionTypeEnum);
-        return false;
+    private boolean isValid(Map<TranscriptionStatusEnum, Set<TranscriptionStatusEnum>> transitionRules,
+                            TranscriptionStatusEnum currentTranscriptionStatus,
+                            TranscriptionStatusEnum desiredTargetTranscriptionStatus,
+                            TranscriptionTypeEnum transcriptionTypeEnum) {
+        boolean isValid = transitionRules.get(currentTranscriptionStatus).contains(desiredTargetTranscriptionStatus);
+        if (!isValid) {
+            log.warn("Unable to go from workflow state {} to {} for type {}",
+                     currentTranscriptionStatus,
+                     desiredTargetTranscriptionStatus,
+                     transcriptionTypeEnum);
+        }
+        return isValid;
     }
 
 }
